Move all Passport.js code to /src/server/passport.js
